+++ packages/patternfly-4/react-charts/src/components/ChartTheme/index.ts
@@ -1,10 +1,2 @@
-export {
-  ChartBaseTheme,
-  ChartDonutTheme,
-  ChartDonutThresholdDynamicTheme,
-  ChartDonutThresholdStaticTheme,
-  ChartDonutUtilizationDynamicTheme,
-  ChartDonutUtilizationStaticTheme,
-  ChartThemeColor,
-  ChartThemeVariant
-} from './ChartTheme';
+export * from './ChartTheme';
+export * from './ChartStyles';