+++ packages/patternfly-4/react-charts/src/components/ChartTheme/styles/common-styles.ts
@@ -1,0 +1,12 @@
+import { global_FontFamily_sans_serif } from '@patternfly/react-tokens';
+
+export const CommonStyles = {
+  label: {
+    fontFamily: global_FontFamily_sans_serif.value,
+    margin: 8
+  },
+  legend: {
+    margin: 16,
+    position: 'right'
+  },
+};