+++ packages/patternfly-4/react-core/src/components/DataList/index.js
@@ -1,9 +1,0 @@
-export { default as DataList } from './DataList';
-export { default as DataListAction } from './DataListAction';
-export { default as DataListCell } from './DataListCell';
-export { default as DataListCheck } from './DataListCheck';
-export { default as DataListItem } from './DataListItem';
-export { default as DataListItemCells } from './DataListItemCells';
-export { default as DataListItemRow } from './DataListItemRow';
-export { default as DataListToggle } from './DataListToggle';
-export { default as DataListContent } from './DataListContent';