+++ packages/patternfly-4/react-docs/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.8.33](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.32...@patternfly/react-docs@4.8.33) (2019-06-25)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.32](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.31...@patternfly/react-docs@4.8.32) (2019-06-25)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.31](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.30...@patternfly/react-docs@4.8.31) (2019-06-25)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.30](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.29...@patternfly/react-docs@4.8.30) (2019-06-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.28...@patternfly/react-docs@4.8.29) (2019-06-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.27...@patternfly/react-docs@4.8.28) (2019-06-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.26...@patternfly/react-docs@4.8.27) (2019-06-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.25...@patternfly/react-docs@4.8.26) (2019-06-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.8.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.24...@patternfly/react-docs@4.8.25) (2019-06-24)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [4.8.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.8.23...@patternfly/react-docs@4.8.24) (2019-06-21)
 
 **Note:** Version bump only for package @patternfly/react-docs