+++ packages/patternfly-4/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,78 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.25.14](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.13...demo-app-ts@1.25.14) (2019-06-25)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.25.13](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.12...demo-app-ts@1.25.13) (2019-06-25)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.25.12](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.11...demo-app-ts@1.25.12) (2019-06-25)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.25.11](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.10...demo-app-ts@1.25.11) (2019-06-24)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.25.10](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.9...demo-app-ts@1.25.10) (2019-06-24)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.25.9](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.8...demo-app-ts@1.25.9) (2019-06-24)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.25.8](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.7...demo-app-ts@1.25.8) (2019-06-24)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.25.7](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.6...demo-app-ts@1.25.7) (2019-06-24)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.25.6](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.5...demo-app-ts@1.25.6) (2019-06-24)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
 ## [1.25.5](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.25.4...demo-app-ts@1.25.5) (2019-06-21)
 
 **Note:** Version bump only for package demo-app-ts